Add files via upload
An Indian startup named 'Foodie' wants to build a conversational bot (chatbot) which can help users discover restaurants across several Indian cities. You have been hired as the lead data scientist for creating this product.
The main purpose of the bot is to help users discover restaurants quickly and efficiently and to provide a good restaurant discovery experience. The project brief provided to you is as follows.The bot takes the following inputs from the user:
City: Take the input from the customer as a text field. For example:
Bot: In which city are you looking for restaurants?
User: anywhere in Delhi